Bullet holes disapearing
There was a bug when I added the enemies that when they dies the bullet holes were still there, so I made the bullet holes store the object they are on and delete when the enemy deletes, I also made them despawn after 10 seconds. I also downloaded a skybox pack.