Enemy states
I added patrol, chase and attack states for the enemy. Now the enemy will check if it can see the player, if it can't then it will 'patrol', so go to random points. Then if it is in sight range of player it will move to player. Then if it is within 'attack' range then it will stand still.